config/config.py
import os
import secrets
from dotenv import load_dotenv

load_dotenv()

class Config:
    SECRET_KEY = os.environ.get('SECRET_KEY') or secrets.token_hex(16)
    SQLALCHEMY_DATABASE_URI = 'sqlite:///database.db'
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    UPLOAD_FOLDER = 'uploads'
    SESSION_PERMANENT = True
    SESSION_TYPE = 'filesystem'
    OPENAI_API_KEY = os.environ.get('OPENAI_API_KEY')