requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
faiss-cpu==1.10.0
filelock==3.18.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.48
langchain-openai==0.1.14
langgraph==0.3.20
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.0
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.24.7
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
xhtml2pdf==0.2.15
docx2txt==0.8
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.1.0
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==78.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
zstandard==0.23.0
